# design/api.yaml
openapi: 3.0.3
info:
  title: Gridow API
  description: 格物（Gridow）物品管理后台服务接口文档，本地数据作为缓存，后台服务作为数据提供者。
  version: 1.0.0
  contact:
    name: Gridow Team
servers:
  - url: https://api.gridow.com/v1
    description: 生产环境
  - url: http://localhost:8080/v1
    description: 开发环境
...
paths:
...
  /spaces/{spaceId}:
    get:
      tags: [空间]
      summary: 获取空间详情
      parameters:
        - $ref: "#/components/parameters/SpaceId"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [空间]
      summary: 更新空间
      parameters:
        - $ref: "#/components/parameters/SpaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpaceUpdateRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [空间]
      summary: 删除空间
      description: 递归删除空间及其所有子空间，解除关联物品的位置绑定
      parameters:
        - $ref: "#/components/parameters/SpaceId"
      responses:
        "204":
          description: 删除成功
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    SpaceId:
      name: spaceId
      in: path
      required: true
      description: 空间ID
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 错误码
        message:
          type: string
          description: 错误信息
...
    Space:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 客厅
        icon:
          type: string
          default: "🏠"
          example: "🏠"
        count:
          type: integer
          description: 空间下物品数量（含子空间）
          default: 0
        parentId:
          type: string
          format: uuid
          nullable: true
          description: 父空间ID，顶级空间为null
        depth:
          type: integer
          description: 空间深度（顶级为0）
          default: 0
        sortOrder:
          type: integer
          default: 0
        photoUri:
          type: string
          default: ""
          description: 空间图片URL
        ownerId:
          type: string
          description: 所有者用户ID
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SpaceUpdateRequest:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        photoUri:
          type: string
        sortOrder:
          type: integer
...