# src/handlers/space.rs
use axum::extract::{Path, Query, State};
use axum::Json;
use serde::Deserialize;
use sqlx::PgPool;
use uuid::Uuid;

use crate::auth::{AuthUser, OptionalAuthUser};
use crate::models::error::AppError;
use crate::models::item::Item;
use crate::models::space::*;
use crate::state::AppState;

async fn check_access(pool: &PgPool, user_id: &str, entity_type: &str, entity_id: &str, owner_id: &str) -> Result<(), AppError> {
    if owner_id == user_id {
        return Ok(());
    }
    let (count,): (i64,) = sqlx::query_as(
        "SELECT COUNT(*) FROM collaborators WHERE entity_type = $1 AND entity_id = $2 AND user_id = $3"
    )
    .bind(entity_type)
    .bind(entity_id)
    .bind(user_id)
    .fetch_one(pool)
    .await
    .map_err(AppError::Database)?;

    if count > 0 {
        return Ok(());
    }
    Err(AppError::Forbidden)
}

pub async fn list_spaces(
    State(state): State<AppState>,
    OptionalAuthUser { user_id }: OptionalAuthUser,
    Query(params): Query<SpaceListParams>,
) -> Result<Json<Vec<Space>>, AppError> {
    let spaces = if let Some(ref uid) = user_id {
        let owner_condition = format!(
            " AND (owner_id = '{}' OR id IN (SELECT entity_id FROM collaborators WHERE entity_type = 'space' AND user_id = '{}'))",
            uid, uid
        );
        if let Some(ref parent_id) = params.parent_id {
            let query = format!(
                "SELECT * FROM spaces WHERE parent_id = $1{} ORDER BY sort_order, name",
                owner_condition
            );
            sqlx::query_as::<_, Space>(&query)
                .bind(parent_id)
                .fetch_all(&state.db)
                .await
                .map_err(AppError::Database)?
        } else {
            let query = format!(
                "SELECT * FROM spaces WHERE parent_id IS NULL{} ORDER BY sort_order, name",
                owner_condition
            );
            sqlx::query_as::<_, Space>(&query)
                .fetch_all(&state.db)
                .await
                .map_err(AppError::Database)?
        }
    } else if let Some(ref parent_id) = params.parent_id {
        sqlx::query_as::<_, Space>("SELECT * FROM spaces WHERE parent_id = $1 ORDER BY sort_order, name")
            .bind(parent_id)
            .fetch_all(&state.db)
            .await
            .map_err(AppError::Database)?
    } else {
        sqlx::query_as::<_, Space>("SELECT * FROM spaces WHERE parent_id IS NULL ORDER BY sort_order, name")
            .fetch_all(&state.db)
            .await
            .map_err(AppError::Database)?
    };

    Ok(Json(spaces))
}

pub async fn create_space(
    State(state): State<AppState>,
    auth: AuthUser,
    Json(req): Json<SpaceCreateRequest>,
) -> Result<(axum::http::StatusCode, Json<Space>), AppError> {
    let id = Uuid::new_v4().to_string();
    let now = chrono::Utc::now()
        .naive_utc()
        .format("%Y-%m-%d %H:%M:%S")
        .to_string();

    let (depth, parent_id) = if let Some(ref pid) = req.parent_id {
        let parent: Space = sqlx::query_as("SELECT * FROM spaces WHERE id = $1")
            .bind(pid)
            .fetch_optional(&state.db)
            .await
            .map_err(AppError::Database)?
            .ok_or(AppError::BadRequest("父空间不存在".to_string()))?;
        (parent.depth + 1, Some(pid.clone()))
    } else {
        (0, None)
    };

    let space = sqlx::query_as::<_, Space>(
        r#"INSERT INTO spaces (id, name, icon, count, parent_id, depth, sort_order, photo_uri, owner_id, created_at, updated_at)
           VALUES ($1, $2, $3, 0, $4, $5, 0, $6, $7, $8, $9)
           RETURNING *"#,
    )
    .bind(&id)
    .bind(&req.name)
    .bind(&req.icon)
    .bind(&parent_id)
    .bind(depth)
    .bind(&req.photo_uri)
    .bind(&auth.user_id)
    .bind(&now)
    .bind(&now)
    .fetch_one(&state.db)
    .await
    .map_err(AppError::Database)?;

    Ok((axum::http::StatusCode::CREATED, Json(space)))
}

pub async fn get_space(
    State(state): State<AppState>,
    Path(space_id): Path<String>,
) -> Result<Json<Space>, AppError> {
    let space = sqlx::query_as::<_, Space>("SELECT * FROM spaces WHERE id = $1")
        .bind(&space_id)
        .fetch_optional(&state.db)
        .await
        .map_err(AppError::Database)?
        .ok_or(AppError::NotFound)?;

    Ok(Json(space))
}

pub async fn update_space(
    State(state): State<AppState>,
    auth: AuthUser,
    Path(space_id): Path<String>,
    Json(req): Json<SpaceUpdateRequest>,
) -> Result<Json<Space>, AppError> {
    let existing = sqlx::query_as::<_, Space>("SELECT * FROM spaces WHERE id = $1")
        .bind(&space_id)
        .fetch_optional(&state.db)
        .await
        .map_err(AppError::Database)?
        .ok_or(AppError::NotFound)?;

    check_access(&state.db, &auth.user_id, "space", &space_id, &existing.owner_id).await?;

    let name = req.name.unwrap_or(existing.name);
    let icon = req.icon.unwrap_or(existing.icon);
    let photo_uri = req.photo_uri.unwrap_or(existing.photo_uri);
    let sort_order = req.sort_order.unwrap_or(existing.sort_order);
    let now = chrono::Utc::now()
        .naive_utc()
        .format("%Y-%m-%d %H:%M:%S")
        .to_string();

    let space = sqlx::query_as::<_, Space>(
        "UPDATE spaces SET name=$1, icon=$2, photo_uri=$3, sort_order=$4, updated_at=$5 WHERE id=$6 RETURNING *",
    )
    .bind(&name)
    .bind(&icon)
    .bind(&photo_uri)
    .bind(sort_order)
    .bind(&now)
    .bind(&space_id)
    .fetch_one(&state.db)
    .await
    .map_err(AppError::Database)?;

    Ok(Json(space))
}

pub async fn delete_space(
    State(state): State<AppState>,
    auth: AuthUser,
    Path(space_id): Path<String>,
) -> Result<axum::http::StatusCode, AppError> {
    let space = sqlx::query_as::<_, Space>("SELECT * FROM spaces WHERE id = $1")
        .bind(&space_id)
        .fetch_optional(&state.db)
        .await
        .map_err(AppError::Database)?
        .ok_or(AppError::NotFound)?;

    if space.owner_id != auth.user_id {
        return Err(AppError::Forbidden);
    }

    delete_space_recursive(&state, &space_id).await?;

    sqlx::query("DELETE FROM collaborators WHERE entity_type = 'space' AND entity_id = $1")
        .bind(&space_id)
        .execute(&state.db)
        .await
        .map_err(AppError::Database)?;

    Ok(axum::http::StatusCode::NO_CONTENT)
}

async fn delete_space_recursive(state: &AppState, space_id: &str) -> Result<(), AppError> {
    let mut stack = vec![space_id.to_string()];

    while let Some(current_id) = stack.pop() {
        let children: Vec<Space> =
            sqlx::query_as("SELECT * FROM spaces WHERE parent_id = $1")
                .bind(&current_id)
                .fetch_all(&state.db)
                .await
                .map_err(AppError::Database)?;

        for child in children {
            stack.push(child.id);
        }

        sqlx::query("UPDATE items SET location_id=NULL, location='' WHERE location_id = $1")
            .bind(&current_id)
            .execute(&state.db)
            .await
            .map_err(AppError::Database)?;

        sqlx::query("DELETE FROM spaces WHERE id = $1")
            .bind(&current_id)
            .execute(&state.db)
            .await
            .map_err(AppError::Database)?;
    }

    Ok(())
}

pub async fn get_space_tree(
    State(state): State<AppState>,
    OptionalAuthUser { user_id }: OptionalAuthUser,
) -> Result<Json<Vec<SpaceNode>>, AppError> {
    let all_spaces: Vec<Space> = if let Some(ref uid) = user_id {
        let query = format!(
            "SELECT * FROM spaces WHERE (owner_id = '{}' OR id IN (SELECT entity_id FROM collaborators WHERE entity_type = 'space' AND user_id = '{}')) ORDER BY sort_order, name",
            uid, uid
        );
        sqlx::query_as(&query)
            .fetch_all(&state.db)
            .await
            .map_err(AppError::Database)?
    } else {
        sqlx::query_as("SELECT * FROM spaces ORDER BY sort_order, name")
            .fetch_all(&state.db)
            .await
            .map_err(AppError::Database)?
    };

    let owner_ids: Vec<String> = all_spaces.iter().map(|s| s.id.clone()).collect();
    let item_locations: Vec<(String, Option<String>)> = if owner_ids.is_empty() {
        vec![]
    } else {
        let placeholders: Vec<String> = (1..=owner_ids.len()).map(|i| format!("${}", i)).collect();
        let mut query_builder = sqlx::QueryBuilder::new(
            format!("SELECT id, location_id FROM items WHERE location_id IN ({})", placeholders.join(","))
        );
        for oid in &owner_ids {
            query_builder.push_bind(oid.clone());
        }
        query_builder
            .build_query_as()
            .fetch_all(&state.db)
            .await
            .map_err(AppError::Database)?
    };

    let mut item_id_map: std::collections::HashMap<String, Vec<String>> = std::collections::HashMap::new();
    for (item_id, location_id) in item_locations {
        if let Some(loc_id) = location_id {
            item_id_map.entry(loc_id).or_default().push(item_id);
        }
    }

    let root_nodes = build_space_tree(&all_spaces, None, &item_id_map);
    Ok(Json(root_nodes))
}

fn build_space_tree(
    spaces: &[Space],
    parent_id: Option<&str>,
    item_id_map: &std::collections::HashMap<String, Vec<String>>,
) -> Vec<SpaceNode> {
    let mut nodes = Vec::new();

    for space in spaces {
        let is_child = match (parent_id, &space.parent_id) {
            (None, None) => true,
            (Some(pid), Some(spid)) => pid == spid,
            _ => false,
        };

        if is_child {
            let children = build_space_tree(spaces, Some(&space.id), item_id_map);

            let item_ids = item_id_map.get(&space.id).cloned().unwrap_or_default();

            let node = SpaceNode {
                id: space.id.clone(),
                name: space.name.clone(),
                icon: space.icon.clone(),
                count: space.count,
                parent_id: space.parent_id.clone(),
                depth: space.depth,
                photo_uri: space.photo_uri.clone(),
                children,
                item_ids,
                owner_id: space.owner_id.clone(),
            };
            nodes.push(node);
        }
    }

    nodes
}

pub async fn get_space_children(
    State(state): State<AppState>,
    Path(space_id): Path<String>,
) -> Result<Json<Vec<Space>>, AppError> {
    let _space = sqlx::query_as::<_, Space>("SELECT * FROM spaces WHERE id = $1")
        .bind(&space_id)
        .fetch_optional(&state.db)
        .await
        .map_err(AppError::Database)?
        .ok_or(AppError::NotFound)?;

    let children =
        sqlx::query_as::<_, Space>("SELECT * FROM spaces WHERE parent_id = $1 ORDER BY sort_order, name")
            .bind(&space_id)
            .fetch_all(&state.db)
            .await
            .map_err(AppError::Database)?;

    Ok(Json(children))
}

pub async fn get_space_items(
    State(state): State<AppState>,
    Path(space_id): Path<String>,
) -> Result<Json<Vec<Item>>, AppError> {
    let _space = sqlx::query_as::<_, Space>("SELECT * FROM spaces WHERE id = $1")
        .bind(&space_id)
        .fetch_optional(&state.db)
        .await
        .map_err(AppError::Database)?
        .ok_or(AppError::NotFound)?;

    let items = sqlx::query_as::<_, Item>("SELECT * FROM items WHERE location_id = $1")
        .bind(&space_id)
        .fetch_all(&state.db)
        .await
        .map_err(AppError::Database)?;

    Ok(Json(items))
}

pub async fn get_space_path(
    State(state): State<AppState>,
    Path(space_id): Path<String>,
) -> Result<Json<SpacePathResponse>, AppError> {
    let segments = get_space_path_segments(&state.db, &space_id).await?;

    let path = segments
        .iter()
        .map(|s| format!("{} {}", s.icon, s.name))
        .collect::<Vec<_>>()
        .join(" > ");

    Ok(Json(SpacePathResponse { path, segments }))
}

pub async fn get_space_path_segments(
    pool: &PgPool,
    space_id: &str,
) -> Result<Vec<SpacePathSegment>, AppError> {
    let mut segments = Vec::new();
    let mut current_id = Some(space_id.to_string());

    while let Some(id) = current_id {
        let space: Space = sqlx::query_as("SELECT * FROM spaces WHERE id = $1")
            .bind(&id)
            .fetch_optional(pool)
            .await
            .map_err(AppError::Database)?
            .ok_or(AppError::NotFound)?;

        segments.push(SpacePathSegment {
            id: space.id.clone(),
            name: space.name.clone(),
            icon: space.icon.clone(),
        });

        current_id = space.parent_id;
    }

    segments.reverse();
    Ok(segments)
}

#[derive(Debug, Deserialize)]
pub struct SpaceListParams {
    parent_id: Option<String>,
}